Protocol errors fixed. get_bit_arr developed.

+++ Ladder.h
@@ -49,11 +49,11 @@ template <>
 class LdVarDevice<IO_IN_AL_GEN>:public LdVar{
 private:
   Device<IO_IN_AL_GEN> *device;
-  int qtDivs, *divs;
+  uint16_t qtDivs, *divs;
   uint8_t *zoneVals, *dominances;
 public:
   LdVarDevice();
-  LdVarDevice(uint8_t id, DeviceBase *baseDev, int qtDivs, int *divs, uint8_t *zoneVals, uint8_t *dominances);
+  LdVarDevice(uint8_t id, DeviceBase *baseDev, uint16_t qtDivs, uint16_t *divs, uint8_t *zoneVals, uint8_t *dominances);
   ~LdVarDevice();
   uint8_t getValue();
 };
@@ -62,4 +62,3 @@ template <>
 uint8_t LdVarDevice<IO_OUT_DG>::getValue();
 template <>
 void LdVarDevice<IO_OUT_DG>::setValue(uint8_t value);
-
